Feature/setup basic layout - basic layout of all pages set up (#8)
Basic layout of all pages set up using Bootsrap and Thymeleaf for page generation.

Navigation is present at the top, with Logo and application name. Example bookmarks are present. At the bottom - basic footer is present with copryright information.

All elements are prestyled with Bootsrap and configured to suit the color scheme, etc.

Page is mobile responsive and navigation will be closable if the screen is too small.